Building client side of project
Installed bower dependencies in public folder and those include: angular, bootstrap, angular-ui-router (routing), angular-strap (display error+success messages), and angular-cookies (to save users). + Setup base for client side for part of application including css for basic styling for search box, buttons, and results to display, etc... and the index.html to display views, and includes scripts from bower components, and public scripts for controllers (some of which will be added in the future).