Setting up Angular side of the project (learning about angular)
Include html views (displayed data on interface). home is what'll be displayed on main page, addSiteis the page displayed for adding data (URL, title, and decription), login holds a login for and register holds a register form for registering new users. + setup angular portion part of project by adding dependecies in app.js, defining routes and controllers in routes_config + started on controllers for each of the states/pages user could be in.